Algorithms #7 Reverse Integer Barry write by C
Result:
1032 / 1032 test cases passed.
Status: Accepted
Runtime: 19 ms

+++ 7_Reverse Integer_Barry.c
@@ -3,7 +3,7 @@
 int reverse(int x)
 {
     /* To avoid overflow while calculate, use double size int to save the input value. */
-	long long out = 0, xx = (long)x;
+    long out = 0, xx = (long)x;
     int MAX_INT = 0x7FFFFFFF, MIN_INT = 0x80000000;
     /* Use while loop to go-through all the input number of bits */
     while(xx)
@@ -14,21 +14,21 @@ int reverse(int x)
     /* Checking if the reverse value is overflow of int(4 bytes) */
     if( out <= MAX_INT && out >= MIN_INT )
     {
-		return (int)out;
-	}
+        return (int)out;
+    }
     else
     {
-		return 0;
-	}
+        return 0;
+    }
 }
 
 int main()
 {
-   int input_num, rev_num;
-   printf("Please input a int value: \n");
-   scanf("%d", &input_num);
-   rev_num = reverse(input_num);
-   printf("The reverse of the input number is: %d\n", rev_num);
-   system("PAUSE");
-   return 0;
+    int input_num, rev_num;
+    printf("Please input a int value: \n");
+    scanf("%d", &input_num);
+    rev_num = reverse(input_num);
+    printf("The reverse of the input number is: %d\n", rev_num);
+    system("PAUSE");
+    return 0;
 }
